articles/models.py
from django.db import models
from django.utils import timezone
from PIL import Image


class Article(models.Model):
    author = models.CharField(max_length=60)
    title = models.CharField(max_length=120)
    content = models.TextField()
    date_posted = models.DateTimeField(default=timezone.now)
    image = models.ImageField(default="default.jpg", upload_to="profile_pics")

    def __str__(self):
        return self.title

    def save(self, *args, **kwargs):
        super().save()

        img = Image.open(self.image.path)

        if img.height > 300 or img.width > 300:
            output_size = (300, 300)
            img.thumbnail(output_size)
            img.save(self.image.path)
